feat(Datamapper): Add Datamapper to the catalog
This commit adds a new `kaoto-patterns.json` asset that gets included in
the Camel Catalog.

This file, in combination with the `camel-component-default.service.ts`
file, allows the user to add new Kaoto Datamapper elements in the
canvas.

This `kaoto-patterns.json` file could also be extended in the future, to
allow users to add their own custom patterns, in a form of a shortcut,
although a better way to do it could be using a kamelet instead.