fix(DataMapper): Provide DnD hint to XPath editor
Currently, there's no hint for the user to drag fields or functions from
the right panel to the source code editor when editing a XPath.

This commit updates the function menu to use the drag/dragging mouse
cursor when interacting with it, and it also adds a question mark icon
to explain to the user they can drag items from the left to the right
panel.

fix: https://github.com/KaotoIO/kaoto-datamapper-integration/issues/180